refactor(tests): migrate mcp_tests_legacy.rs to per-domain test modules
Distributes the 100 tests from the 2,981-line mcp_tests_legacy.rs into:
- mcp_tests/tool_tests.rs    (49 tests)
- mcp_tests/error_tests.rs   (23 tests)
- mcp_tests/prompt_tests.rs  (20 tests)
- mcp_tests/resource_tests.rs (8 tests)

Removes the legacy file and its Cargo.toml entry.
Total test count is unchanged.
