+++ apps/things3-cli/Cargo.toml
@@ -77,11 +77,6 @@ path = "tests/mcp_testing_framework_tests.rs"
 required-features = ["mcp-server"]
 
 [[test]]
-name = "mcp_tests_legacy"
-path = "tests/mcp_tests_legacy.rs"
-required-features = ["mcp-server"]
-
-[[test]]
 name = "mcp_write_operations_tests"
 path = "tests/mcp_write_operations_tests.rs"
 required-features = ["mcp-server"]
